requirements.txt
boto3==1.34.131
urllib3==1.25.4
botocore==1.34.131
langchain 
streamlit==1.22.0
langchain_aws
langchain_community
streamlit-chat==0.0.2.2
streamlit-extras==0.2.7
python-dotenv